[WIP] Estable inversion and corrects RNIP for plane interfaces
The inversion algorithm is working for plane interfaces models (3 and 4) and dynamic ray tracing is getting the correct RNIP parameters. The problem occurs because the velocity model interpolation close to interfaces is affecting the dynamic ray system that depends of the partial velocity derivatives. Also, the runge-kutta interval should be equal to dt, instead of 2dt. And you should pay atention if you are using Offset or Half-offset coordinate.

Making this corrections, the algorithm works properly for plane interfaces because velocity interpolation does not affects ray dynamic system so much.